webkit/LayoutTests/fast/canvas/radialGradient-negative-radii-expected.txt

This test checks that createRadialGradient throws an INDEX_SIZE_ERR exception if either of the specified radii (or both) are negative, as per section 5 of the HTML Canvas 2D Context specification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context.createRadialGradient(0, 0, -1.0, 0, 0, 1.0) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS context.createRadialGradient(0, 0, 1.0, 0, 0, -1.0) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.
PASS context.createRadialGradient(0, 0, -1.0, 0, 0, -1.0) threw exception Error: INDEX_SIZE_ERR: DOM Exception 1.

TEST COMPLETE

